+++ packages/taro-cli/src/rn.js
@@ -4,6 +4,7 @@ const {performance} = require('perf_hooks')
 const chokidar = require('chokidar')
 const chalk = require('chalk')
 const vfs = require('vinyl-fs')
+const ejs = require('ejs')
 const Vinyl = require('vinyl')
 const through2 = require('through2')
 const babel = require('babel-core')
@@ -20,6 +21,7 @@ const npmProcess = require('./util/npm')
 const CONFIG = require('./config')
 const babylonConfig = require('./config/babylon')
 const AstConvert = require('./util/astConvert')
+const {getPkgVersion} = require('./util')
 
 const appPath = process.cwd()
 const projectConfig = require(path.join(appPath, Util.PROJECT_CONFIG))(_.merge)
@@ -31,7 +33,9 @@ const entryFilePath = Util.resolveScriptPath(path.join(sourceDir, CONFIG.ENTRY))
 const entryFileName = path.basename(entryFilePath)
 const pluginsConfig = projectConfig.plugins || {}
 
-const isBuildingStyles = {}
+const pkgPath = path.join(__dirname, './rn/pkg')
+
+let isBuildingStyles = {}
 const styleDenpendencyTree = {}
 
 const reactImportDefaultName = 'React'
@@ -62,9 +66,18 @@ const PACKAGES = {
 }
 
 function parseJSCode (code, filePath) {
-  const ast = babel.transform(code, {
-    parserOpts: babylonConfig
-  }).ast
+  let ast
+  try {
+    ast = babel.transform(code, {
+      parserOpts: babylonConfig,
+      plugins: ['babel-plugin-transform-jsx-stylesheet']
+    }).ast
+  } catch (e) {
+    if (e.name === 'ReferenceError') {
+      console.log(chalk.yellow('警告：请安装 npm 包 babel-plugin-transform-jsx-stylesheet'))
+    }
+    throw e
+  }
   const styleFiles = []
   let pages = [] // app.js 里面的config 配置里面的 pages
   let iconPaths = [] // app.js 里面的config 配置里面的需要引入的 iconPath
@@ -76,7 +89,6 @@ function parseJSCode (code, filePath) {
   let hasAppExportDefault
   let componentClassName
   let classRenderReturnJSX
-  let importStyleName
 
   traverse(ast, {
     ImportDeclaration (astPath) {
@@ -95,12 +107,8 @@ function parseJSCode (code, filePath) {
           if (styleFiles.indexOf(stylePath) < 0) {
             styleFiles.push(stylePath)
           }
-          importStyleName = _.camelCase(`${basename}_styles`)
-          const importSpecifiers = [t.importDefaultSpecifier(t.identifier(importStyleName))]
-          astPath.replaceWith(t.importDeclaration(
-            importSpecifiers,
-            t.stringLiteral(`${path.dirname(value)}/${basename}_styles`))
-          )
+          // index.css -> index_styles
+          astPath.node.source = t.stringLiteral(`${path.dirname(value)}/${basename}_styles`)
         }
         return
       }
@@ -193,6 +201,10 @@ function parseJSCode (code, filePath) {
                     if (node.key.name === 'navigationBarBackgroundColor' || node.key.value === 'navigationBarBackgroundColor') {
                       navigationOptions['headerStyle'] = {backgroundColor: node.value.value}
                     }
+                    // 开启下拉刷新
+                    if (node.key.name === 'enablePullDownRefresh' || node.key.value === 'enablePullDownRefresh') {
+                      navigationOptions['enablePullDownRefresh'] = node.value.value
+                    }
                   }
                 })
                 astPath.replaceWith(t.objectProperty(
@@ -211,7 +223,7 @@ function parseJSCode (code, filePath) {
                       node.key.value === 'selectedIconPath'
                     ) {
                       if (typeof value !== 'string') return
-                      let iconName = value.replace(/\/|\./g, '')
+                      let iconName = _.camelCase(value.split('/'))
                       iconPaths.push(value)
                       astPath.insertAfter(t.objectProperty(
                         t.identifier(node.key.name || node.key.value),
@@ -242,6 +254,10 @@ function parseJSCode (code, filePath) {
               if (node.key.name === 'navigationBarBackgroundColor' || node.key.value === 'navigationBarBackgroundColor') {
                 navigationOptions['headerStyle'] = {backgroundColor: node.value.value}
               }
+              // 开启下拉刷新
+              if (node.key.name === 'enablePullDownRefresh' || node.key.value === 'enablePullDownRefresh') {
+                navigationOptions['enablePullDownRefresh'] = node.value.value
+              }
             }
           })
           astPath.replaceWith(t.classProperty(
@@ -250,38 +266,6 @@ function parseJSCode (code, filePath) {
           ))
           astPath.node.static = 'true'
         }
-      }
-    },
-    // 转换 className 和 id
-    JSXElement (astPath) {
-      const node = astPath.node
-      const openingElement = node.openingElement
-      if (openingElement && openingElement.attributes.length) {
-        const attributes = openingElement.attributes
-        const newAttributes = []
-        let styleAttrs = [] // classNames 值
-        attributes.forEach(attr => {
-          const name = attr.name
-          if (name.name === 'className' || name.name === 'id') {
-            if (attr.value) {
-              styleAttrs = styleAttrs.concat(attr.value.value.split(' '))
-            }
-          } else {
-            newAttributes.push(attr)
-          }
-        })
-        if (styleAttrs.length) {
-          styleAttrs = _.uniq(styleAttrs)
-          // 合成 RN style
-          const styleArr = styleAttrs.map(item => {
-            const styleName = `${importStyleName}.${item}`
-            return t.identifier(styleName)
-          })
-          newAttributes.push(
-            t.jSXAttribute(t.jSXIdentifier('style'), t.jSXExpressionContainer(t.arrayExpression(styleArr)))
-          )
-        }
-        openingElement.attributes = newAttributes
       }
     },
     // 获取 classRenderReturnJSX
@@ -417,7 +401,7 @@ function parseJSCode (code, filePath) {
           // 注入 import page from 'XXX'
           pages.forEach(item => {
             const pagePath = item.startsWith('/') ? item : `/${item}`
-            const screenName = pagePath.replace(/\//g, '')
+            const screenName = _.camelCase(pagePath.split('/'), {pascalCase: true})
             const importScreen = template(
               `import ${screenName} from '.${pagePath}'`,
               babylonConfig
@@ -426,7 +410,7 @@ function parseJSCode (code, filePath) {
           })
           iconPaths.forEach(item => {
             const iconPath = item.startsWith('/') ? item : `/${item}`
-            const iconName = iconPath.replace(/\/|\./g, '')
+            const iconName = _.camelCase(iconPath.split('/'))
             const importIcon = template(
               `import ${iconName} from '.${iconPath}'`,
               babylonConfig
@@ -435,10 +419,10 @@ function parseJSCode (code, filePath) {
           })
           // Taro.initRouter  生成 RootStack
           const routerPages = pages
-            .map(pageItem => {
-              const pageName = pageItem.startsWith('/') ? pageItem : `/${pageItem}`
-              const screenName = pageName.replace(/\//g, '')
-              return `['${pageItem}',${screenName}]`
+            .map(item => {
+              const pagePath = item.startsWith('/') ? item : `/${item}`
+              const screenName = _.camelCase(pagePath.split('/'), {pascalCase: true})
+              return `['${item}',${screenName}]`
             })
             .join(',')
           node.body.push(template(
@@ -489,6 +473,7 @@ function compileDepStyles (filePath, styleFiles) {
   return Promise.all(styleFiles.map(async p => {
     const filePath = path.join(p)
     const fileExt = path.extname(filePath)
+    Util.printLog(Util.pocessTypeEnum.COMPILE, _.camelCase(fileExt).toUpperCase(), filePath)
     const pluginName = Util.FILE_PROCESSOR_MAP[fileExt]
     if (pluginName) {
       return npmProcess.callPlugin(pluginName, null, filePath, pluginsConfig[pluginName] || {})
@@ -525,6 +510,8 @@ function compileDepStyles (filePath, styleFiles) {
     } catch (err) {
       console.log(err)
     }
+  }).catch((e) => {
+    throw new Error(e)
   })
 }
 
@@ -543,6 +530,7 @@ function buildTemp () {
           return cb()
         }
         if (Util.REG_SCRIPT.test(filePath)) {
+          Util.printLog(Util.pocessTypeEnum.COMPILE, 'JS', filePath)
           let transformResult = parseJSCode(content, filePath)
           const jsCode = transformResult.code
           const styleFiles = transformResult.styleFiles
@@ -560,11 +548,14 @@ function buildTemp () {
             }
           }, null, 2))
         })
-        // 后期可以改为模版实现
-        const pkgObj = Object.assign({}, {name: projectConfig.projectName}, require('./rn/pkg'))
+        // .temp 下的 package.json
+        const pkgContent = ejs.render(fs.readFileSync(pkgPath, 'utf-8'), {
+          projectName: projectConfig.projectName,
+          version: getPkgVersion()
+        })
         const pkg = new Vinyl({
           path: 'package.json',
-          contents: Buffer.from(JSON.stringify(pkgObj, null, 2))
+          contents: Buffer.from(pkgContent)
         })
         // Copy bin/crna-entry.js ?
         const crnaEntryPath = path.join(path.dirname(npmProcess.resolveNpmSync('@tarojs/rn-runner')), 'src/bin/crna-entry.js')
@@ -617,6 +608,7 @@ async function buildDist ({watch}) {
 }
 
 async function processFiles (filePath) {
+  isBuildingStyles = {} // 清空
   // 后期可以优化，不编译全部
   let t0 = performance.now()
   await buildTemp()