+++ packages/taro-cli/src/weapp.js
@@ -26,7 +26,8 @@ const defaultBabelConfig = require('./config/babel')
 const defaultTSConfig = require('./config/tsconfig.json')
 
 const appPath = process.cwd()
-const projectConfig = require(path.join(appPath, Util.PROJECT_CONFIG))(_.merge)
+const configDir = path.join(appPath, Util.PROJECT_CONFIG)
+const projectConfig = require(configDir)(_.merge)
 const sourceDirName = projectConfig.sourceRoot || CONFIG.SOURCE_DIR
 const outputDirName = projectConfig.outputRoot || CONFIG.OUTPUT_DIR
 const sourceDir = path.join(appPath, sourceDirName)
@@ -36,6 +37,12 @@ const entryFileName = path.basename(entryFilePath)
 const outputEntryFilePath = path.join(outputDir, entryFileName)
 
 const pluginsConfig = projectConfig.plugins || {}
+const weappConf = projectConfig.weapp || {}
+const weappNpmConfig = Object.assign({
+  name: CONFIG.NPM_DIR,
+  dir: null
+}, weappConf.npm)
+const appOutput = typeof weappConf.appOutput === 'boolean' ? weappConf.appOutput : true
 
 const notExistNpmList = []
 const taroJsFramework = '@tarojs/taro'
@@ -60,13 +67,21 @@ const PARSE_AST_TYPE = {
   NORMAL: 'NORMAL'
 }
 
+const DEVICE_RATIO = 'deviceRatio'
+
 const isWindows = os.platform() === 'win32'
 
 function getExactedNpmFilePath (npmName, filePath) {
   try {
-    const npmInfo = resolveNpmFilesPath(npmName, isProduction)
+    const npmInfo = resolveNpmFilesPath(npmName, isProduction, weappNpmConfig)
     const npmInfoMainPath = npmInfo.main
-    const outputNpmPath = npmInfoMainPath.replace('node_modules', path.join(outputDirName, CONFIG.NPM_DIR))
+    let outputNpmPath
+    if (!weappNpmConfig.dir) {
+      outputNpmPath = npmInfoMainPath.replace('node_modules', path.join(outputDirName, weappNpmConfig.name))
+    } else {
+      const npmFilePath = npmInfoMainPath.replace(/(.*)node_modules/, '')
+      outputNpmPath = path.join(path.resolve(configDir, '..', weappNpmConfig.dir), weappNpmConfig.name, npmFilePath)
+    }
     const relativePath = path.relative(filePath, outputNpmPath)
     return Util.promoteRelativePath(relativePath)
   } catch (err) {
@@ -117,11 +132,16 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath) {
   traverse(ast, {
     ClassDeclaration (astPath) {
       const node = astPath.node
+      let hasCreateData = false
       if (node.superClass) {
-        if (node.superClass.name === 'Component' ||
-        node.superClass.name === 'BaseComponent' ||
-        (node.superClass.type === 'MemberExpression' &&
-        node.superClass.object.name === taroImportDefaultName)) {
+        astPath.traverse({
+          ClassMethod (astPath) {
+            if (astPath.get('key').isIdentifier({ name: '_createData' })) {
+              hasCreateData = true
+            }
+          }
+        })
+        if (hasCreateData) {
           needExportDefault = true
           astPath.traverse({
             ClassMethod (astPath) {
@@ -166,10 +186,15 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath) {
     ClassExpression (astPath) {
       const node = astPath.node
       if (node.superClass) {
-        if (node.superClass.name === 'Component' ||
-        node.superClass.name === 'BaseComponent' ||
-        (node.superClass.type === 'MemberExpression' &&
-        node.superClass.object.name === taroImportDefaultName)) {
+        let hasCreateData = false
+        astPath.traverse({
+          ClassMethod (astPath) {
+            if (astPath.get('key').isIdentifier({ name: '_createData' })) {
+              hasCreateData = true
+            }
+          }
+        })
+        if (hasCreateData) {
           needExportDefault = true
           if (node.id === null) {
             componentClassName = '_TaroComponentClass'
@@ -317,22 +342,28 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath) {
         (declaration.type === 'ClassDeclaration' || declaration.type === 'ClassExpression')
       ) {
         const superClass = declaration.superClass
-        if (superClass &&
-          (superClass.name === 'Component' ||
-          superClass.name === 'BaseComponent' ||
-          (superClass.type === 'MemberExpression' &&
-          superClass.object.name === taroImportDefaultName))) {
-          needExportDefault = true
-          if (declaration.id === null) {
-            componentClassName = '_TaroComponentClass'
-          } else if (declaration.id.name === 'App') {
-            componentClassName = '_App'
-          } else {
-            componentClassName = declaration.id.name
+        if (superClass) {
+          let hasCreateData = false
+          astPath.traverse({
+            ClassMethod (astPath) {
+              if (astPath.get('key').isIdentifier({ name: '_createData' })) {
+                hasCreateData = true
+              }
+            }
+          })
+          if (hasCreateData) {
+            needExportDefault = true
+            if (declaration.id === null) {
+              componentClassName = '_TaroComponentClass'
+            } else if (declaration.id.name === 'App') {
+              componentClassName = '_App'
+            } else {
+              componentClassName = declaration.id.name
+            }
+            const isClassDcl = declaration.type === 'ClassDeclaration'
+            const classDclProps = [t.identifier(componentClassName), superClass, declaration.body, declaration.decorators || []]
+            astPath.replaceWith(isClassDcl ? t.classDeclaration.apply(null, classDclProps) : t.classExpression.apply(null, classDclProps))
           }
-          const isClassDcl = declaration.type === 'ClassDeclaration'
-          const classDclProps = [t.identifier(componentClassName), superClass, declaration.body, declaration.decorators || []]
-          astPath.replaceWith(isClassDcl ? t.classDeclaration.apply(null, classDclProps) : t.classExpression.apply(null, classDclProps))
         }
       } else if (declaration.type === 'CallExpression') {
         const callee = declaration.callee
@@ -575,19 +606,22 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath) {
           node.body.push(exportDefault)
         }
         const taroWeappFrameworkPath = getExactedNpmFilePath(taroWeappFramework, filePath)
-        let insert
         switch (type) {
           case PARSE_AST_TYPE.ENTRY:
-            insert = template(`App(require('${taroWeappFrameworkPath}').default.createApp(${exportVariableName}))`, babylonConfig)()
-            node.body.push(insert)
+            const pxTransformConfig = {
+              designWidth: projectConfig.designWidth || 750,
+            }
+            if (projectConfig.hasOwnProperty(DEVICE_RATIO)) {
+              pxTransformConfig[DEVICE_RATIO] = projectConfig.deviceRatio
+            }
+            node.body.push(template(`App(require('${taroWeappFrameworkPath}').default.createApp(${exportVariableName}))`, babylonConfig)())
+            node.body.push(template(`Taro.initPxTransform(${JSON.stringify(pxTransformConfig)})`, babylonConfig)())
             break
           case PARSE_AST_TYPE.PAGE:
-            insert = template(`Page(require('${taroWeappFrameworkPath}').default.createComponent(${exportVariableName}, true))`, babylonConfig)()
-            node.body.push(insert)
+            node.body.push(template(`Page(require('${taroWeappFrameworkPath}').default.createComponent(${exportVariableName}, true))`, babylonConfig)())
             break
           case PARSE_AST_TYPE.COMPONENT:
-            insert = template(`Component(require('${taroWeappFrameworkPath}').default.createComponent(${exportVariableName}))`, babylonConfig)()
-            node.body.push(insert)
+            node.body.push(template(`Component(require('${taroWeappFrameworkPath}').default.createComponent(${exportVariableName}))`, babylonConfig)())
             break
           default:
             break
@@ -596,7 +630,7 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath) {
     }
   })
   return {
-    code: generate(ast).code,
+    code: unescape(generate(ast).code.replace(/\\u/g, '%u')),
     styleFiles,
     scriptFiles,
     jsonFiles,
@@ -649,7 +683,7 @@ function convertArrayToAstExpression (arr) {
       return t.nullLiteral()
     }
     if (typeof value === 'object') {
-      return convertObjectToAstExpression(value)
+      return t.objectExpression(convertObjectToAstExpression(value))
     }
   })
 }
@@ -730,10 +764,12 @@ async function buildEntry () {
         }
       }
     }
-    fs.writeFileSync(path.join(outputDir, 'app.json'), JSON.stringify(res.configObj, null, 2))
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '入口配置', `${outputDirName}/app.json`)
-    fs.writeFileSync(path.join(outputDir, 'app.js'), resCode)
-    Util.printLog(Util.pocessTypeEnum.GENERATE, '入口文件', `${outputDirName}/app.js`)
+    if (appOutput) {
+      fs.writeFileSync(path.join(outputDir, 'app.json'), JSON.stringify(res.configObj, null, 2))
+      Util.printLog(Util.pocessTypeEnum.GENERATE, '入口配置', `${outputDirName}/app.json`)
+      fs.writeFileSync(path.join(outputDir, 'app.js'), resCode)
+      Util.printLog(Util.pocessTypeEnum.GENERATE, '入口文件', `${outputDirName}/app.js`)
+    }
     const fileDep = dependencyTree[entryFilePath] || {}
     // 编译依赖的脚本文件
     if (Util.isDifferentArray(fileDep['script'], res.scriptFiles)) {
@@ -950,7 +986,6 @@ async function buildSinglePage (page) {
 }
 
 async function processStyleWithPostCSS (styleObj) {
-  const weappConf = projectConfig.weapp || {}
   const useModuleConf = weappConf.module || {}
   const customPostcssConf = useModuleConf.postcss || {}
   const customPxtransformConf = customPostcssConf.pxtransform || {}
@@ -960,7 +995,6 @@ async function processStyleWithPostCSS (styleObj) {
     platform: 'weapp'
   }
 
-  const DEVICE_RATIO = 'deviceRatio'
   if (projectConfig.hasOwnProperty(DEVICE_RATIO)) {
     postcssPxtransformOption[DEVICE_RATIO] = projectConfig.deviceRatio
   }
@@ -1182,7 +1216,6 @@ function compileDepScripts (scriptFiles) {
   scriptFiles.forEach(async item => {
     if (path.isAbsolute(item)) {
       const outputItem = item.replace(path.join(sourceDir), path.join(outputDir)).replace(path.extname(item), '.js')
-      const weappConf = projectConfig.weapp || {}
       const useCompileConf = Object.assign({}, weappConf.compile)
       const compileExclude = useCompileConf.exclude || []
       let isInCompileExclude = false