+++ packages/taro-h5/src/api/system/index.js
@@ -1,0 +1,9 @@
+import { getSystemInfo, getSystemInfoSync } from './info'
+import { getNetworkType, onNetworkStatusChange } from './network'
+
+export {
+  getSystemInfo,
+  getSystemInfoSync,
+  getNetworkType,
+  onNetworkStatusChange
+}