+++ packages/taro-h5/src/native-api.js
@@ -1,9 +1,10 @@
-import { onAndSyncApis, noPromiseApis, otherApis } from '@tarojs/taro'
+import { onAndSyncApis, noPromiseApis, otherApis, initPxTransform } from '@tarojs/taro'
 import { createSelectorQuery } from './api/createSelectorQuery'
 import request from './api/request'
 import * as storage from './api/storage'
 import * as interactive from './api/interactive'
 import webSocket from './api/webSocket'
+import * as system from './api/system'
 
 function processApis (taro) {
   const weApis = Object.assign({ }, onAndSyncApis, noPromiseApis, otherApis)
@@ -14,9 +15,16 @@ function processApis (taro) {
   })
 }
 
+function pxTransform (size) {
+  const { designWidth } = this.config
+  return Math.ceil((parseInt(size, 10) / 40 * 640 / designWidth) * 10000) / 10000 + 'rem'
+}
+
 export default function initNativeApi (taro) {
   processApis(taro)
   taro.request = request
   taro.createSelectorQuery = createSelectorQuery
-  Object.assign(taro, storage, interactive, webSocket)
+  taro.initPxTransform = initPxTransform.bind(taro)
+  taro.pxTransform = pxTransform.bind(taro)
+  Object.assign(taro, storage, interactive, webSocket, system)
 }