+++ packages/taro-transformer-wx/src/index.ts
@@ -2,12 +2,11 @@ import traverse, { Binding, NodePath } from 'babel-traverse'
 import generate from 'babel-generator'
 import { Transformer } from './class'
 import { prettyPrint } from 'html'
-import { setting, findFirstIdentifierFromMemberExpression, isContainJSXElement } from './utils'
+import { setting, findFirstIdentifierFromMemberExpression, isContainJSXElement, codeFrameError } from './utils'
 import * as t from 'babel-types'
-import { DEFAULT_Component_SET, INTERNAL_SAFE_GET, TARO_PACKAGE_NAME, ASYNC_PACKAGE_NAME, REDUX_PACKAGE_NAME, INTERNAL_DYNAMIC, IMAGE_COMPONENTS, INTERNAL_INLINE_STYLE } from './constant'
+import { DEFAULT_Component_SET, INTERNAL_SAFE_GET, TARO_PACKAGE_NAME, ASYNC_PACKAGE_NAME, REDUX_PACKAGE_NAME, INTERNAL_DYNAMIC, IMAGE_COMPONENTS, INTERNAL_INLINE_STYLE, THIRD_PARTY_COMPONENTS } from './constant'
 import { transform as parse } from 'babel-core'
 import * as ts from 'typescript'
-import { remove } from 'lodash'
 const template = require('babel-template')
 
 export interface Options {
@@ -48,7 +47,7 @@ function resetTSClassProperty (body: (t.ClassMethod | t.ClassProperty)[]) {
   for (const method of body) {
     if (t.isClassMethod(method) && method.kind === 'constructor') {
       if (t.isBlockStatement(method.body)) {
-        for (const statement of method.body.body) {
+        method.body.body = method.body.body.filter(statement => {
           if (t.isExpressionStatement(statement) && t.isAssignmentExpression(statement.expression)) {
             const expr = statement.expression
             const { left, right } = expr
@@ -62,21 +61,31 @@ function resetTSClassProperty (body: (t.ClassMethod | t.ClassProperty)[]) {
                 ||
                 (left.property.name === 'config' && t.isObjectExpression(right))
               ) {
-                body.push(
-                  t.classProperty(left.property, right)
-                )
-                remove(method.body.body, statement)
+                const classProp = t.classProperty(left.property, right)
+                body.push(classProp)
+                handleThirdPartyComponent(classProp)
+                return false
               }
             }
           }
-        }
+          return true
+        })
       }
     }
   }
 }
 
+function findDeclarationScope (path: NodePath<t.Node>, id: t.Identifier) {
+  const scopePath = path.findParent(p => !!p.scope.getOwnBindingIdentifier(id.name))
+  if (scopePath) {
+    return scopePath
+  }
+  throw codeFrameError(path.node, '该引用从未被定义')
+}
+
 function buildFullPathThisPropsRef (id: t.Identifier, memberIds: string[], path: NodePath<t.Node>) {
-  const binding = path.scope.getOwnBinding(id.name)
+  const scopePath = findDeclarationScope(path, id)
+  const binding = scopePath.scope.getOwnBinding(id.name)
   if (binding) {
     const bindingPath = binding.path
     if (bindingPath.isVariableDeclarator()) {
@@ -90,6 +99,30 @@ function buildFullPathThisPropsRef (id: t.Identifier, memberIds: string[], path:
         }
         if (dclInitIds[0] === 'this' && dclInitIds[1] === 'props') {
           return template(dclInitIds.concat(memberIds).join('.'))().expression
+        }
+      }
+    }
+  }
+}
+
+function handleThirdPartyComponent (expr: t.ClassMethod | t.ClassProperty) {
+  if (t.isClassProperty(expr) && expr.key.name === 'config' && t.isObjectExpression(expr.value)) {
+    const properties = expr.value.properties
+    for (const prop of properties) {
+      if (
+        t.isObjectProperty(prop) &&
+        (t.isIdentifier(prop.key, { name: 'usingComponents' }) || t.isStringLiteral(prop.key, { value: 'usingComponents' })) &&
+        t.isObjectExpression(prop.value)
+      ) {
+        for (const value of prop.value.properties) {
+          if (t.isObjectProperty(value)) {
+            if (t.isStringLiteral(value.key)) {
+              THIRD_PARTY_COMPONENTS.add(value.key.value)
+            }
+            if (t.isIdentifier(value.key)) {
+              THIRD_PARTY_COMPONENTS.add(value.key.name)
+            }
+          }
         }
       }
     }
@@ -139,7 +172,8 @@ export default function transform (options: Options): TransformResult {
         'objectRestSpread',
         'decorators'
       ] as any[]
-    }
+    },
+    plugins: [[require('babel-plugin-danger-remove-unused-import'), { ignore: ['Taro'] }]]
   }).ast as t.File
   if (options.isNormal) {
     return { ast } as any
@@ -168,11 +202,10 @@ export default function transform (options: Options): TransformResult {
       if (isContainJSXElement(path)) {
         return
       }
-
       if (callee.isReferencedMemberExpression()) {
         const id = findFirstIdentifierFromMemberExpression(callee.node)
         const calleeIds = getIdsFromMemberProps(callee.node)
-        if (t.isIdentifier(id)) {
+        if (t.isIdentifier(id) && id.name.startsWith('on')) {
           const fullPath = buildFullPathThisPropsRef(id, calleeIds, path)
           if (fullPath) {
             path.replaceWith(
@@ -188,7 +221,7 @@ export default function transform (options: Options): TransformResult {
       if (callee.isReferencedIdentifier()) {
         const id = callee.node
         const ids = [id.name]
-        if (t.isIdentifier(id)) {
+        if (t.isIdentifier(id) && id.name.startsWith('on')) {
           const fullPath = buildFullPathThisPropsRef(id, ids, path)
           if (fullPath) {
             path.replaceWith(
@@ -272,9 +305,12 @@ export default function transform (options: Options): TransformResult {
       const expr = value.expression as any
       const exprPath = path.get('value.expression')
       if (!t.isBinaryExpression(expr, { operator: '+' }) && !t.isLiteral(expr) && name.name === 'style') {
-        exprPath.replaceWith(
-          t.callExpression(t.identifier(INTERNAL_INLINE_STYLE), [expr])
-        )
+        const jsxID = path.findParent(p => p.isJSXOpeningElement()).get('name')
+        if (jsxID && jsxID.isJSXIdentifier() && DEFAULT_Component_SET.has(jsxID.node.name)) {
+          exprPath.replaceWith(
+            t.callExpression(t.identifier(INTERNAL_INLINE_STYLE), [expr])
+          )
+        }
       }
 
       if (name.name.startsWith('on')) {
@@ -331,7 +367,7 @@ export default function transform (options: Options): TransformResult {
           const name = path.node.imported.name
           DEFAULT_Component_SET.has(name) || names.push(name)
           if (source === TARO_PACKAGE_NAME && name === 'Component') {
-            path.node.local = t.identifier('BaseComponent')
+            path.node.local = t.identifier('__BaseComponent')
           }
         }
       })
@@ -341,8 +377,10 @@ export default function transform (options: Options): TransformResult {
   if (!mainClass) {
     throw new Error('未找到 Taro.Component 的类定义')
   }
+
+  mainClass.node.body.body.forEach(handleThirdPartyComponent)
   const storeBinding = mainClass.scope.getBinding(storeName)
-  mainClass.scope.rename('Component', 'BaseComponent')
+  mainClass.scope.rename('Component', '__BaseComponent')
   if (storeBinding) {
     const statementPath = storeBinding.path.getStatementParent()
     if (statementPath) {
@@ -359,7 +397,9 @@ export default function transform (options: Options): TransformResult {
   }
   resetTSClassProperty(mainClass.node.body.body)
   if (options.isApp) {
-    renderMethod.remove()
+    renderMethod.replaceWith(
+      t.classMethod('method', t.identifier('_createData'), [], t.blockStatement([]))
+    )
     return { ast } as TransformResult
   }
   result = new Transformer(mainClass, options.sourcePath).result