+++ packages/taro-weapp/src/native-api.js
@@ -1,4 +1,9 @@
-import { onAndSyncApis, noPromiseApis, otherApis } from '@tarojs/taro'
+import {
+  onAndSyncApis,
+  noPromiseApis,
+  otherApis,
+  initPxTransform
+} from '@tarojs/taro'
 
 const RequestQueue = {
   MAX_REQUEST: 5,
@@ -106,9 +111,20 @@ function processApis (taro) {
   })
 }
 
+function pxTransform (size) {
+  const { designWidth, deviceRatio } = this.config
+  if (!(designWidth in deviceRatio)) {
+    throw new Error(`deviceRatio 配置中不存在 ${designWidth} 的设置！`)
+    return
+  }
+  return parseInt(size, 10) / deviceRatio[designWidth] + 'rpx'
+}
+
 export default function initNativeApi (taro) {
   processApis(taro)
   taro.request = request
   taro.getCurrentPages = getCurrentPages
   taro.getApp = getApp
+  taro.initPxTransform = initPxTransform.bind(taro)
+  taro.pxTransform = pxTransform.bind(taro)
 }